Introduce log entry state machine
The state machine for a log entry is stored in the raft_entry_ref for
that entry, which is a sort of "witness" that the entry is alive. This
is more convenient than storing the sm inside or next to raft_entry
itself, since we also use raft_entry to mean other things than "an
active entry in the in-memory log".

This commit also includes a little internal refactor to pass around the
raft_entry_ref to more places so that sm_init and sm_fini (and soon,
sm_move and sm_fail) can get to it.

Signed-off-by: Cole Miller <cole.miller@canonical.com>